requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy>=2.0.35
pymysql>=1.1.0
python-jose[cryptography]>=3.3.0
passlib==1.7.4
bcrypt==3.2.0
python-multipart>=0.0.9
pydantic[email]>=2.7.0
python-dotenv>=1.0.1
alembic>=1.13.1
httpx>=0.27.0